Change App component to class componet and initialize constructor
