+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"D:\\Microverse\\React\\calculatorReact\\src\\index.js":"1","D:\\Microverse\\React\\calculatorReact\\src\\components\\App.js":"2","D:\\Microverse\\React\\calculatorReact\\src\\components\\Display.js":"3","D:\\Microverse\\React\\calculatorReact\\src\\components\\ButtonPanel.js":"4","D:\\Microverse\\React\\calculatorReact\\src\\logic\\calculate.js":"5","D:\\Microverse\\React\\calculatorReact\\src\\components\\Button.js":"6","D:\\Microverse\\React\\calculatorReact\\src\\logic\\operate.js":"7","D:\\Microverse\\React\\calculatorReact\\src\\buttons\\buttons.js":"8"},{"size":267,"mtime":1610360678107,"results":"9","hashOfConfig":"10"},{"size":821,"mtime":1610457769121,"results":"11","hashOfConfig":"10"},{"size":293,"mtime":1610437672155,"results":"12","hashOfConfig":"10"},{"size":424,"mtime":1610457835152,"results":"13","hashOfConfig":"10"},{"size":1292,"mtime":1610437672155,"results":"14","hashOfConfig":"10"},{"size":328,"mtime":1610457664080,"results":"15","hashOfConfig":"10"},{"size":590,"mtime":1610456771128,"results":"16","hashOfConfig":"10"},{"size":165,"mtime":1610457397522,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1ddvgeo",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},"D:\\Microverse\\React\\calculatorReact\\src\\index.js",[],"D:\\Microverse\\React\\calculatorReact\\src\\components\\App.js",[],"D:\\Microverse\\React\\calculatorReact\\src\\components\\Display.js",[],"D:\\Microverse\\React\\calculatorReact\\src\\components\\ButtonPanel.js",[],"D:\\Microverse\\React\\calculatorReact\\src\\logic\\calculate.js",[],[],"D:\\Microverse\\React\\calculatorReact\\src\\components\\Button.js",[],"D:\\Microverse\\React\\calculatorReact\\src\\logic\\operate.js",[],"D:\\Microverse\\React\\calculatorReact\\src\\buttons\\buttons.js",[]]
