+++ src/buttons/buttons.js
@@ -1,0 +1,9 @@
+const buttons = [
+  ['AC', '+/-', '%', '÷'],
+  ['7', '8', '9', 'X'],
+  ['4', '5', '6', '-'],
+  ['1', '2', '3', '+'],
+  ['0', '.', '='],
+];
+
+export default buttons;
